Add map helper for direct JavaScript function application. (#185)
* feat: Add map helper for direct JS function application.
* chore: Refactor spread column initialization.
* feat: Force serialization errors in map helper.
* test: Add tests for expression helpers.
* docs: Document the map expression helper.
