+++ src/engine/spread.js
@@ -6,7 +6,7 @@ import toArray from '../util/to-array';
 export default function(table, { names, exprs, ops = [] }, options = {}) {
   if (names.length === 0) return table;
 
-  // ignore as if there are multiple field names
+  // ignore 'as' if there are multiple field names
   const as = (names.length === 1 && options.as) || [];
   const drop = options.drop == null ? true : !!options.drop;
   const limit = options.limit == null
@@ -43,13 +43,16 @@ export default function(table, { names, exprs, ops = [] }, options = {}) {
 function spread(table, get, limit) {
   const nrows = table.totalRows();
   const columns = [];
+  let j = -1;
 
   table.scan((row, data) => {
     const values = toArray(get(row, data));
     const n = Math.min(values.length, limit);
+    while (++j < n) {
+      columns.push(Array(nrows).fill(NULL));
+    }
     for (let i = 0; i < n; ++i) {
-      const column = columns[i] || (columns[i] = Array(nrows).fill(NULL));
-      column[row] = values[i];
+      columns[i][row] = values[i];
     }
   });
